added traits
