+++ Utilis/Communicator.php
@@ -1,0 +1,9 @@
+<? php 
+
+namespace Bookstore\Utils;
+
+trait Communicator {
+    public function sign(){
+        echo 'Signing the contract - Communicator';
+    } 
+}
