+++ Utilis/Contract.php
@@ -1,0 +1,9 @@
+<? php 
+
+namespace Bookstore\Utils;
+
+trait Contract {
+    public function sign(){
+        echo 'Signing the contract';
+    } 
+}
